+++ libs/containers/abstract/GuildChannel.lua
@@ -1,4 +1,5 @@
 local json = require('json')
+local enums = require('enums')
 local class = require('class')
 local Channel = require('containers/abstract/Channel')
 local PermissionOverwrite = require('containers/PermissionOverwrite')
@@ -7,6 +8,11 @@ local Cache = require('iterables/Cache')
 
 local isInstance = class.isInstance
 local classes = class.classes
+local channelType = enums.channelType
+
+local insert, sort = table.insert, table.sort
+local min, max, floor = math.min, math.max, math.floor
+local huge = math.huge
 
 local GuildChannel, get = class('GuildChannel', Channel)
 
@@ -35,13 +41,102 @@ function GuildChannel:setName(name)
 	return self:_modify({name = name or json.null})
 end
 
+local function sorter(a, b)
+	if a.position == b.position then
+		return tonumber(a.id) < tonumber(b.id)
+	else
+		return a.position < b.position
+	end
+end
+
+local function getSortedChannels(self)
+
+	local channels
+	if self._type == channelType.text then
+		channels = self._parent._text_channels
+	else
+		channels = self._parent._voice_channels
+	end
+
+	local ret = {}
+	for channel in channels:iter() do
+		insert(ret, {id = channel._id, position = channel._position})
+	end
+	sort(ret, sorter)
+
+	return ret
+
+end
+
+local function setSortedChannels(self, channels)
+	local data, err = self.client._api:modifyGuildChannelPositions(self._parent._id, channels)
+	if data then
+		return true
+	else
+		return false, err
+	end
+end
+
 --[[
-@method setPosition
-@param position: number
+@method moveUp
+@param [n]: number
 @ret boolean
 ]]
-function GuildChannel:setPosition(position)
-	return self:_modify({position = position or json.null})
+function GuildChannel:moveUp(n)
+
+	n = tonumber(n) or 1
+	if n < 0 then
+		return self:moveDown(-n)
+	end
+
+	local channels = getSortedChannels(self)
+
+	local new = huge
+	for i = #channels - 1, 0, -1 do
+		local v = channels[i + 1]
+		if v.id == self._id then
+			new = max(0, i - floor(n))
+			v.position = new
+		elseif i >= new then
+			v.position = i + 1
+		else
+			v.position = i
+		end
+	end
+
+	return setSortedChannels(self, channels)
+
+end
+
+--[[
+@method moveDown
+@param [n]: number
+@ret boolean
+]]
+function GuildChannel:moveDown(n)
+
+	n = tonumber(n) or 1
+	if n < 0 then
+		return self:moveUp(-n)
+	end
+
+	local channels = getSortedChannels(self)
+
+	local new = -huge
+	for i = 0, #channels - 1 do
+		local v = channels[i + 1]
+		if v.id == self._id then
+			new = min(i + floor(n), #channels - 1)
+			v.position = new
+		elseif i <= new then
+			v.position = i - 1
+		else
+			v.position = i
+		end
+	end
+
+	return setSortedChannels(self, channels)
+
 end
 
 --[[